pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "visual-volatility-intelligence"
version = "0.1.0"
description = "Market regime forecasting from implied volatility surfaces"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "scipy>=1.12",
  "matplotlib>=3.8",
  "Pillow>=10.0",
  "scikit-learn>=1.4",
  "torch>=2.2",
  "torchvision>=0.17",
  "fastapi>=0.115",
  "uvicorn>=0.30",
  "pydantic>=2.8",
  "python-dateutil>=2.9"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "jupyter>=1.0",
  "httpx>=0.27"
]

[project.scripts]
vvi-run = "vvi.pipeline:main"
vvi-api = "vvi.api.main:run"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["vvi*"]
